Rallye-Server/requirements.txt
Django~=3.1.2
Pillow
#psycopg2
psycopg2-binary
pytz

djangorestframework~=3.12.1